+++ eTickets/Data/ViewModels/ShoppingCartVM.cs
@@ -10,6 +10,6 @@ namespace eTickets.Data.ViewModels
     {
         public ShoppingCart ShoppingCart { get; set; }
 
-        public double ShoppingCartTotal { get; set; }
+        public decimal ShoppingCartTotal { get; set; }
     }
 }
